Updated priced job estimatives.
Using the shop closest to the center of the map instead of the first one
added to belief base. For the battery estimative we are now multiplying
it by the number of tasks in the job.

This means that we are not going to take most of the jobs for now, which
is okay because they said they will up the reward of jobs, and when we
add assembling capabilities we should start taking more jobs.